pyproject.toml
[project]
name = "agent-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.36.0",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
]
